modify

+++ scripts/envs/develop/DeleteSecretDev.ps1
@@ -1,30 +1,35 @@
 ### Import Variables
-Invoke-Expression (Get-Content ".\VariablesKvDev.ps1" -Raw)
+Invoke-Expression (Get-Content ".\VariablesDev.ps1" -Raw)
 
 ### Functions
 # Remove Secrets
 function RemoveKeyVaultSecret {
   param (
+    [string]$Environment,
     [string]$FilePath,
     [string[]]$CsvFileNameList
   )
 
   foreach ($CsvFileName in $CsvFileNameList) {
-    $PathToCsvFile = $FilePath + "\" + $CsvFileName
-    $SecretList = Import-Csv -Path $PathToCsvFile -Encoding UTF8
+    $PathToCsvFile = Join-Path -Path $FilePath -ChildPath $CsvFileName
+    $CsvFileList = Import-Csv -Path $PathToCsvFile -Encoding UTF8
 
-    foreach ($Secret in $SecretList) {
-      $SecretName = $Secret.SecretName
-      $KeyVaultName = $Secret.KeyVaultName
-      Remove-AzKeyVaultSecret -VaultName $KeyVaultName -Name $SecretName -PassThru -Force
+    foreach ($CsvFile in $CsvFileList) {
+      if ($CsvFile.IAM -match "シークレット") {
+        $SecretName = $CsvFile.ParamName
+        $KeyVaultName = $CsvFile.KeyVaultName.Replace("○○", $Environment.ToLower())
+        Remove-AzKeyVaultSecret -VaultName $KeyVaultName -Name $SecretName -PassThru -Force
+      }
     }
-    # Wait for 15 seconds
-    Start-Sleep -Seconds 15
+      # Wait for 15 seconds
+      Start-Sleep -Seconds 15
 
-    foreach ($Secret in $SecretList) {
-      $SecretName = $Secret.SecretName
-      $KeyVaultName = $Secret.KeyVaultName
-      Remove-AzKeyVaultSecret -VaultName $KeyVaultName -Name $SecretName -InRemovedState -PassThru -Force
+    foreach ($CsvFile in $CsvFileList) {
+      if ($CsvFile.IAM -match "シークレット") {
+        $SecretName = $CsvFile.ParamName
+        $KeyVaultName = $CsvFile.KeyVaultName.Replace("○○", $Environment.ToLower())
+        Remove-AzKeyVaultSecret -VaultName $KeyVaultName -Name $SecretName -InRemovedState -PassThru -Force
+      }
     }
   }
 }
@@ -33,10 +38,10 @@ function RemoveKeyVaultSecret {
 Connect-AzAccount -Subscription $SubscriptionId -Tenant $TenantId
 
 # Delete Secrets
-RemoveKeyVaultSecret -FilePath $FilePath -CsvFileNameList $CsvFileNameList
+RemoveKeyVaultSecret -Environment $Environment -FilePath $FilePath -CsvFileNameList $CsvFileNameList
 $Reply = Read-Host "Do you want to execute this procedure? (y/n)"
 if ($Reply -eq "y") {
-  RemoveKeyVaultSecret -FilePath $FilePath -CsvFileNameList $CsvFileNameWithExpDateList
+  RemoveKeyVaultSecret -Environment $Environment -FilePath $FilePath -CsvFileNameList $CsvFileNameWithExpDateList
 } else {
   Write-Host "The procedure was canceled."
 }
